test: move tests out of __tests__ folder
